CP-54351 move lock_queue from /opt
